[Refactor] useScroll() 커스텀 훅을 하위 컴포넌트로 이동 #437
